Refactored skills. Fixed bugs
