+++ configs/tasks/rearrange/place.yaml
@@ -29,6 +29,7 @@ TASK:
     SUCCESS_MEASURE: "place_success"
     SUCCESS_REWARD: 10.0
     SLACK_REWARD: 0.0
+    END_ON_SUCCESS: True
 
     # If true, does not care about navigability or collisions with objects when spawning
     # robot
