+++ configs/tasks/rearrange/set_table.yaml
@@ -1,8 +1,12 @@
 GYM:
-    AUTO_NAME: "SetTable"
-    CLASS_NAME: "RearrangeRLEnv"
-    OBS_KEYS: ['robot_head_depth', 'obj_start_sensor', 'obj_start_gps_compass', 'joint', 'is_holding', 'obj_goal_sensor', 'obj_goal_gps_compass']
-
+    AUTO_NAME: NavPick
+    CLASS_NAME: RearrangeRLEnv
+    OBS_KEYS:
+    - robot_head_depth
+    - obj_start_sensor
+    - obj_start_gps_compass
+    - joint
+    - is_holding
 ENVIRONMENT:
     MAX_EPISODE_STEPS: 4500
 DATASET:
@@ -12,12 +16,11 @@ DATASET:
     SCENES_DIR: data/replica_cad/
 TASK:
     TYPE: RearrangeCompositeTask-v0
-
-    REWARD_MEASURE: "move_obj_reward"
-    SUCCESS_MEASURE: "composite_success"
+    REWARD_MEASURE: move_obj_reward
+    SUCCESS_MEASURE: composite_success
     SUCCESS_REWARD: 100.0
     SLACK_REWARD: -0.01
-
+    END_ON_SUCCESS: true
     MAX_COLLISIONS: -1.0
     COUNT_OBJ_COLLISIONS: true
     COUNT_ROBOT_OBJ_COLLS: false
@@ -67,23 +70,15 @@ TASK:
     ROBOT_FORCE:
         TYPE: RobotForce
         MIN_FORCE: 20.0
-    ROBOT_COLLS:
-        TYPE: RobotCollisions
-    OBJECT_TO_GOAL_DISTANCE:
-        TYPE: ObjectToGoalDistance
-    END_EFFECTOR_TO_OBJECT_DISTANCE:
-        TYPE: EndEffectorToObjectDistance
-    COMPOSITE_SUCCESS:
-        TYPE: CompositeSuccess
-    NUM_STEPS:
-        TYPE: NumStepsMeasure
     MEASUREMENTS:
     - OBJECT_TO_GOAL_DISTANCE
     - ROBOT_FORCE
     - FORCE_TERMINATE
     - ROBOT_COLLS
     - END_EFFECTOR_TO_OBJECT_DISTANCE
+    - DOES_WANT_TERMINATE
     - COMPOSITE_SUCCESS
+    - COMPOSITE_BAD_CALLED_TERMINATE
     - NUM_STEPS
     - DID_VIOLATE_HOLD_CONSTRAINT
     - MOVE_OBJECTS_REWARD
@@ -112,8 +107,6 @@ TASK:
     - ARM_ACTION
     - BASE_VELOCITY
     - REARRANGE_STOP
-    DID_VIOLATE_HOLD_CONSTRAINT:
-        TYPE: DidViolateHoldConstraintMeasure
     USING_SUBTASKS: false
     MOVE_OBJECTS_REWARD:
         TYPE: MoveObjectsReward
@@ -122,7 +115,7 @@ TASK:
         SINGLE_REARRANGE_REWARD: 1.0
         DIST_REWARD: 1.0
         CONSTRAINT_VIOLATE_PEN: 10.0
-        FORCE_PEN: 0.001
+        FORCE_PEN: 0.0
         MAX_FORCE_PEN: 1.0
         FORCE_END_PEN: 10.0
     TARGET_START_GPS_COMPASS_SENSOR:
@@ -143,6 +136,7 @@ TASK:
         SUCCESS_ANGLE_DIST: 0.7
         MUST_LOOK_AT_TARG: true
         HEURISTIC_STOP: false
+        MUST_CALL_STOP: false
 SIMULATOR:
     ACTION_SPACE_CONFIG: v0
     GRASP_IMPULSE: 1000.0
