+++ habitat/config/default.py
@@ -462,6 +462,7 @@ _C.TASK.REARRANGE_NAV_TO_OBJ_REWARD.FORCE_END_PEN = 10.0
 _C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS = CN()
 _C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS.TYPE = "NavToObjSuccess"
 _C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS.MUST_LOOK_AT_TARG = True
+_C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS.MUST_CALL_STOP = True
 # Distance in radians.
 _C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS.SUCCESS_ANGLE_DIST = 0.15
 _C.TASK.REARRANGE_NAV_TO_OBJ_SUCCESS.HEURISTIC_STOP = False
@@ -564,6 +565,10 @@ _C.TASK.COMPOSITE_REWARD = CN()
 _C.TASK.COMPOSITE_REWARD.TYPE = "CompositeReward"
 _C.TASK.COMPOSITE_REWARD.STAGE_COMPLETE_REWARD = 10.0
 _C.TASK.COMPOSITE_REWARD.SUCCESS_REWARD = 10.0
+_C.TASK.DOES_WANT_TERMINATE = CN()
+_C.TASK.DOES_WANT_TERMINATE.TYPE = "DoesWantTerminate"
+_C.TASK.COMPOSITE_BAD_CALLED_TERMINATE = CN()
+_C.TASK.COMPOSITE_BAD_CALLED_TERMINATE.TYPE = "CompositeBadCalledTerminate"
 # -----------------------------------------------------------------------------
 # -----------------------------------------------------------------------------
 # # EQA TASK
