+++ habitat/tasks/rearrange/sub_tasks/nav_to_obj_sensors.py
@@ -490,11 +490,12 @@ class NavToObjSuccess(GeoMeasure):
 
         called_stop = self.does_action_want_stop(task, observations)
 
-        if called_stop:
-            if self._end_on_stop:
-                task.should_end = True
-        else:
-            self._metric = False
+        if self._config.MUST_CALL_STOP:
+            if called_stop:
+                if self._end_on_stop:
+                    task.should_end = True
+            else:
+                self._metric = False
 
     def does_action_want_stop(self, task, obs):
         if self._config.HEURISTIC_STOP:
