+++ habitat_baselines/config/rearrange/ddppo_close_fridge.yaml
@@ -25,9 +25,6 @@ EVAL_KEYS_TO_INCLUDE_IN_NAME: ['reward', 'force', 'success']
 
 BASE_TASK_CONFIG_PATH: configs/tasks/rearrange/close_fridge.yaml
 ENV_NAME: "RearrangeRLEnv"
-EVAL:
-  SPLIT: "eval"
-
 
 RL:
   POLICY:
