+++ habitat_baselines/config/rearrange/ddppo_nav_to_obj.yaml
@@ -25,9 +25,6 @@ EVAL_KEYS_TO_INCLUDE_IN_NAME: ['reward', 'force', 'success']
 
 BASE_TASK_CONFIG_PATH: configs/tasks/rearrange/nav_to_obj.yaml
 ENV_NAME: "RearrangeRLEnv"
-EVAL:
-  SPLIT: "eval"
-
 
 RL:
   POLICY:
