+++ habitat_baselines/config/rearrange/ddppo_pick.yaml
@@ -25,9 +25,6 @@ FORCE_TORCH_SINGLE_THREADED: True
 EVAL_KEYS_TO_INCLUDE_IN_NAME: ['reward', 'force', 'success']
 ENV_NAME: "RearrangeRLEnv"
 
-EVAL:
-  SPLIT: "eval"
-
 RL:
   POLICY:
       name: "PointNavResNetPolicy"
