+++ habitat_baselines/config/rearrange/hab/tp_srl.yaml
@@ -40,7 +40,9 @@ TASK_CONFIG:
       - PLACE_SUCCESS
       - PLACE_REWARD
       - COMPOSITE_NODE_IDX
+      - DOES_WANT_TERMINATE
       - COMPOSITE_SUCCESS
+      - COMPOSITE_BAD_CALLED_TERMINATE
       - COMPOSITE_REWARD
       - NUM_STEPS
       - DID_VIOLATE_HOLD_CONSTRAINT
@@ -74,7 +76,7 @@ NUM_CHECKPOINTS: 20
 FORCE_TORCH_SINGLE_THREADED: True
 EVAL_KEYS_TO_INCLUDE_IN_NAME: ['composite_reward', 'force', 'composite_success']
 EVAL:
-  SPLIT: "eval"
+  SPLIT: "val"
   USE_CKPT_CONFIG: False
   SHOULD_LOAD_CKPT: False
