+++ habitat_baselines/config/rearrange/spap_pick.yaml
@@ -19,9 +19,6 @@ SENSE_PLAN_ACT:
   GRASP_GEN_IS_VERBOSE: True
   IK_DIST_THRESH: 0.1
 
-EVAL:
-  SPLIT: "eval"
-
 TASK_CONFIG:
   SUCCESS_REWARD: 2000.0
   END_ON_SUCCESS: False
