+++ habitat_baselines/config/rearrange/spap_reach_state.yaml
@@ -1,7 +1,5 @@
 BASE_TASK_CONFIG_PATH: configs/tasks/rearrange/reach_state.yaml
 
-EVAL:
-  SPLIT: "eval"
 SENSE_PLAN_ACT:
   VERBOSE: True
   RUN_FREQ: 4
