+++ habitat_baselines/rl/hrl/skills/pick.py
@@ -1,0 +1,57 @@
+import torch
+
+from habitat.tasks.rearrange.rearrange_sensors import (
+    IsHoldingSensor,
+    RelativeRestingPositionSensor,
+)
+from habitat_baselines.rl.hrl.skills.nn_skill import NnSkillPolicy
+
+
+class PickSkillPolicy(NnSkillPolicy):
+    def _is_skill_done(
+        self,
+        observations,
+        rnn_hidden_states,
+        prev_actions,
+        masks,
+    ) -> torch.BoolTensor:
+        # Is the agent holding the object and is the end-effector at the
+        # resting position?
+        rel_resting_pos = torch.norm(
+            observations[RelativeRestingPositionSensor.cls_uuid], dim=-1
+        )
+        is_within_thresh = rel_resting_pos < self._config.AT_RESTING_THRESHOLD
+        is_holding = observations[IsHoldingSensor.cls_uuid].view(-1)
+        return (is_holding * is_within_thresh).type(torch.bool)
+
+    def _parse_skill_arg(self, skill_arg):
+        self._internal_log(f"Parsing skill argument {skill_arg}")
+        return int(skill_arg[0].split("|")[1])
+
+    def _mask_pick(self, action, observations):
+        # Mask out the release if the object is already held.
+        is_holding = observations[IsHoldingSensor.cls_uuid].view(-1)
+        for i in torch.nonzero(is_holding):
+            # Do not release the object once it is held
+            action[i, self._ac_start + self._ac_len - 1] = 1.0
+        return action
+
+    def _internal_act(
+        self,
+        observations,
+        rnn_hidden_states,
+        prev_actions,
+        masks,
+        cur_batch_idx,
+        deterministic=False,
+    ):
+        action, hxs = super()._internal_act(
+            observations,
+            rnn_hidden_states,
+            prev_actions,
+            masks,
+            cur_batch_idx,
+            deterministic,
+        )
+        action = self._mask_pick(action, observations)
+        return action, hxs
