+++ habitat_baselines/rl/hrl/skills/wait.py
@@ -1,0 +1,43 @@
+from typing import Any
+
+import gym.spaces as spaces
+import torch
+
+from habitat_baselines.rl.hrl.skills.skill import SkillPolicy
+
+
+class WaitSkillPolicy(SkillPolicy):
+    def __init__(
+        self,
+        config,
+        action_space: spaces.Space,
+        batch_size,
+    ):
+        super().__init__(config, action_space, batch_size, True)
+        self._wait_time = -1
+
+    def _parse_skill_arg(self, skill_arg: str) -> Any:
+        self._wait_time = int(skill_arg[0])
+        self._internal_log(f"Requested wait time {self._wait_time}")
+
+    def _is_skill_done(
+        self,
+        observations,
+        rnn_hidden_states,
+        prev_actions,
+        masks,
+    ) -> torch.BoolTensor:
+        assert self._wait_time > 0
+        return self._cur_skill_step >= self._wait_time
+
+    def _internal_act(
+        self,
+        observations,
+        rnn_hidden_states,
+        prev_actions,
+        masks,
+        cur_batch_idx,
+        deterministic=False,
+    ):
+        action = torch.zeros(prev_actions.shape, device=prev_actions.device)
+        return action, rnn_hidden_states
